requirements.txt
pytest==7.4.0
selenium==4.10.0
Faker==18.13.0
allure-pytest==2.13.2
pytest-xdist==3.3.1
requests==2.31.0
mariadb==1.1.7
python-dotenv==1.0.0